New Updated.

+++ app/src/main/java/com/example/siddhesh/texttospeech/MainActivity.java
@@ -1,25 +1,33 @@
 package com.example.siddhesh.texttospeech;
 
 import android.content.ContentValues;
+import android.content.Intent;
+import android.content.SharedPreferences;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.graphics.Color;
 import android.speech.tts.TextToSpeech;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.widget.PopupMenu;
+import android.support.v7.widget.Toolbar;
+import android.text.InputType;
 import android.util.Log;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.AutoCompleteTextView;
 import android.widget.Button;
+import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import java.util.Locale;
 
 public class MainActivity extends AppCompatActivity implements TextToSpeech.OnInitListener{
-
+    ImageView his,menus;
+    public static final String DEFAULT="N/A";
     private int result=0;
     private TextToSpeech tts;
     private Button btnSpeak;
@@ -27,19 +35,27 @@ public class MainActivity extends AppCompatActivity implements TextToSpeech.OnIn
     SQLiteDatabase sqLiteDatabase;
     DBhelper dBhelper;
     String[] history;
-    TextView tv;
+    Toolbar toolbar;
     ArrayAdapter<String> adapter;
+    SharedPreferences sharedPreferences;
+    SharedPreferences.Editor editor;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        sharedPreferences = getSharedPreferences("MyData",MODE_PRIVATE);
+        editor = sharedPreferences.edit();
+        toolbar = (Toolbar)findViewById(R.id.tool);
+        his = (ImageView)findViewById(R.id.history_icon);
+        menus = (ImageView)findViewById(R.id.menuic);
+        //setSupportActionBar(toolbar);
         tts = new TextToSpeech(this, this);
         btnSpeak = (Button)findViewById(R.id.btnSpeak);
         txtText = (AutoCompleteTextView) findViewById(R.id.txtText);
-        tv = (TextView)findViewById(R.id.heading);
+        txtText.requestFocus();
         dBhelper = new DBhelper(this);
-        setCustomAdapter();
+            setCustomAdapter();
         //button on click event
         btnSpeak.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -49,13 +65,35 @@ public class MainActivity extends AppCompatActivity implements TextToSpeech.OnIn
                 setCustomAdapter();
             }
         });
-        tv.setOnClickListener(new View.OnClickListener() {
+        menus.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                sqLiteDatabase = dBhelper.getWritableDatabase();
-                sqLiteDatabase.execSQL("delete from HISTORY");
-                Toast.makeText(MainActivity.this, "History Cleared.", Toast.LENGTH_SHORT).show();
-                setCustomAdapter();
+                PopupMenu popupMenu = new PopupMenu(MainActivity.this, v);
+                popupMenu.inflate(R.menu.menu_main);
+                popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
+                    @Override
+                    public boolean onMenuItemClick(MenuItem item) {
+                        if(!item.getTitle().equals("Exit")) {
+                            sqLiteDatabase = dBhelper.getWritableDatabase();
+                            sqLiteDatabase.execSQL("delete from HISTORY");
+                            editor.putString("status","true");
+                            editor.commit();
+                            Toast.makeText(MainActivity.this, "History Cleared.", Toast.LENGTH_SHORT).show();
+                            setCustomAdapter();
+                            return true;
+                        } else {
+                            System.exit(1);
+                        }
+                        return false;
+                    }
+                });
+                popupMenu.show();
+            }
+        });
+        his.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                startActivity(new Intent(MainActivity.this,History.class));
             }
         });
     }
@@ -108,6 +146,8 @@ public class MainActivity extends AppCompatActivity implements TextToSpeech.OnIn
         }
     }
     public void insert() {
+
+        if(!isPresent(txtText.getText().toString())){
         try {
             long id = -1;
             sqLiteDatabase = dBhelper.getWritableDatabase();
@@ -117,10 +157,13 @@ public class MainActivity extends AppCompatActivity implements TextToSpeech.OnIn
             if (id < 0) {
                 Toast.makeText(this, "Failed to add Record.", Toast.LENGTH_LONG).show();
             } else {
+                editor.putString("status","false");
+                editor.commit();
                 //Toast.makeText(this, "Record added successfully.", Toast.LENGTH_LONG).show();
             }
         } catch (Exception e) {
             Toast.makeText(this, "" + e.toString(), Toast.LENGTH_LONG).show();
+        }
         }
     }
 
@@ -140,6 +183,10 @@ public class MainActivity extends AppCompatActivity implements TextToSpeech.OnIn
         }
     }
     public void setCustomAdapter(){
+        if(sharedPreferences.getString("status",DEFAULT).equals("true")) {
+            txtText.setAdapter(null);
+            return;
+        }
         offline();
         adapter = new ArrayAdapter<String>(getApplicationContext(), android.R.layout.simple_list_item_1, history) {
             @Override
@@ -150,6 +197,18 @@ public class MainActivity extends AppCompatActivity implements TextToSpeech.OnIn
                 return textView;
             }
         };
+
         txtText.setAdapter(adapter);
     }
+    public boolean isPresent(String str){
+        if(!sharedPreferences.getString("status",DEFAULT).equals("true")){
+        for(int i=0;i<history.length;i++){
+            if(history[i].equals(str)){
+                return true;
+            }
+        }
+        }
+        return false;
+    }
+
 }
